Spring Boot JPA Join Mapping

+++ Spring Data/Spring-Boot-Mappings/Spring-Boot-Mapping-OneToOne/src/main/java/com/mylearning/springbootmappingonetoone/repository/LaptopRepository.java
@@ -3,5 +3,5 @@ package com.mylearning.springbootmappingonetoone.repository;
 import com.mylearning.springbootmappingonetoone.entity.Laptop;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LaptopRepository extends JpaRepository<Laptop, Long> {
+public interface LaptopRepository extends JpaRepository<Laptop, Integer> {
 }
